% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalc.R
\name{pcalc}
\alias{pcalc}
\title{Calculate PC reaction}
\usage{
pcalc(v = 10, rxn = NULL, initialConc = NULL, finalConc = NULL, pError = 0)
}
\arguments{
\item{v}{The final volume desired for each PC reaction. Numeric.}

\item{rxn}{The number of total reactions. Numeric.}

\item{initialConc}{Data Frame containing the initial concentrations of each component.}

\item{finalConc}{Data Frame containing the final concentrations of each component.}

\item{pError}{Pipetting error to consider for each reaction.}
}
\description{
Returns a efficient protocol for pc reaction setup.
}
