% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitPCR.R
\name{splitPCR}
\alias{splitPCR}
\title{Calculate PC reaction}
\usage{
splitPCR(
  pcr = NULL,
  nVars = c(2, 2, 2),
  vars = c("MM", "DNA", "FPrimer", "RPrimer"),
  levels = c(0, 1, 2, 2)
)
}
\arguments{
\item{vars}{Variable PCR components. Character vector.}

\item{nvars}{The number of variable. Numeric vector.}
}
\description{
Returns a efficient protocol for pc reaction setup.
}
